Splitting llvm::findScalarElement into separate file
Makes dxilconv.dll 121kB smaller (x86 release, compiler optimization for size).
