+++ lib/Analysis/CMakeLists.txt
@@ -70,6 +70,7 @@ add_llvm_library(LLVMAnalysis
   ScopedNoAliasAA.cpp
   ValueTracking.cpp
   VectorUtils.cpp
+  VectorUtils2.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm/Analysis
